avant final

+++ lib/src/classes/passes_temp.dart
@@ -3,7 +3,7 @@ import 'package:flutter_application/src/models/MunitionReferenceEnum.dart';
 class PassesTemp {
   int id;
   MunitionReferenceEnum munitionReference;
-  int profondeurSonde, gradientMag, profondeurSecurisee, coteSecurisee;
+  double profondeurSonde, gradientMag, profondeurSecurisee, coteSecurisee;
 
   PassesTemp(
       {required this.id,
